ORM/SQLALCHEMY/requirements.txt
# Install sqlaclhemy
sqlalchemy

# Python dotenv
python-dotenv

# Install psycopg2-binary
psycopg2-binary

Faker